Removed page for LogOut
